修正資料表欄位

+++ database/migrations/2026_08_29_161204_create_movies_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');                           // 電影名稱
             $table->string('director')->nullable();            // 導演（可為空）
-            $table->integer('release_year');      // 上映年份 (例如: 2024)
+            $table->unsignedSmallInteger('release_year');      // 上映年份 (例如: 2024)
             $table->string('genre')->default('未分類');         // 電影類型（第一階段先用字串存）
             $table->decimal('rating', 3, 1)->default(0.0);     // 評分 (例如: 8.5，範圍 0.0 ~ 10.0)
             $table->text('description')->nullable();           // 劇情簡介
